+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/layer/impl/DividerLineLayer.java
@@ -5,6 +5,9 @@ import dev.xylonity.tooltipoverhaul.client.layer.LayerDepth;
 import dev.xylonity.tooltipoverhaul.client.render.TooltipContext;
 import dev.xylonity.tooltipoverhaul.client.util.Constants;
 import dev.xylonity.tooltipoverhaul.client.util.RenderUtils;
+import dev.xylonity.tooltipoverhaul.client.util.TextUtils;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.phys.Vec2;
 
 public interface DividerLineLayer extends ITooltipLayer {
@@ -23,16 +26,24 @@ public interface DividerLineLayer extends ITooltipLayer {
             int x = (int) context.getTooltipPosition().x + context.getPaddingX();
 
             if (hasIcon) {
-                y += Constants.ICON_SIZE + Constants.SEPARATION_TITLE_ICON;
+                int dividerLinetopPadding = Constants.getDividerLineTopPadding(context);
+                y += Constants.getIconSize(context) + dividerLinetopPadding;
+                if (dividerLinetopPadding > 1) {
+                    y += Constants.getDividerLineHeight(context);
+                }
+
             }
             else {
-                int extraY = context.getComponents().get(0).getHeight();
+                int titleHeight = context.getComponents().get(0).getHeight();
+                y += titleHeight;
+
+                // If there is a rating text present, adds its height
                 if (hasRating) {
-                    y += extraY * 2 + context.getPaddingY();
+                    Component rating = TextUtils.getRatingText(context);
+                    y += ClientTooltipComponent.create(rating.getVisualOrderText()).getHeight();
                 }
-                else {
-                    y += extraY + context.getPaddingY();
-                }
+
+                y += Constants.getDividerLineTopPadding(context);
 
             }