+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/util/Constants.java
@@ -1,10 +1,31 @@
 package dev.xylonity.tooltipoverhaul.client.util;
 
-public final class Constants {
+import dev.xylonity.tooltipoverhaul.client.render.TooltipContext;
 
-    public static int ICON_SIZE = 22;
-    public static int ICON_SIZE_PADDING = ICON_SIZE - 10;
-    public static int SEPARATION_TITLE_ICON = 3;
-    public static int DIVIDER_LINE_FULL_PADDING = 6;
+public class Constants {
+
+    public static int getDividerLineTopPadding(TooltipContext context) {
+        return Math.max(context.getPaddingY(), context.hasIcon() ? 1 : 0) + PositionUtils.getDividerLineTopPadding(context);
+    }
+
+    public static int getDividerLineBottomPadding(TooltipContext context) {
+        return Math.max(context.getPaddingY(), 1) + PositionUtils.getDividerLineBottomPadding(context);
+    }
+
+    public static int getDividerLineHeight(TooltipContext context) {
+        return 1;
+    }
+
+    public static int getDividerLineFullPadding(TooltipContext context) {
+        return getDividerLineBottomPadding(context) + getDividerLineHeight(context) + getDividerLineTopPadding(context);
+    }
+
+    public static int getIconSize(TooltipContext context) {
+        return 22;
+    }
+
+    public static int getIconTitleSeparation(TooltipContext context) {
+        return 3;
+    }
 
 }