+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/util/PositionUtils.java
@@ -16,4 +16,12 @@ public class PositionUtils {
         return Optional.ofNullable(context.getFrameData()).map(CustomFrameData::getTooltipPositionY).orElse(TooltipsConfig.TOOLTIP_POSITION_Y);
     }
 
+    public static int getDividerLineTopPadding(TooltipContext context) {
+        return Optional.ofNullable(context.getFrameData()).map(CustomFrameData::getDividerLineTopPadding).orElse(TooltipsConfig.DIVIDER_LINE_TOP_PADDING);
+    }
+
+    public static int getDividerLineBottomPadding(TooltipContext context) {
+        return Optional.ofNullable(context.getFrameData()).map(CustomFrameData::getDividerLineBottomPadding).orElse(TooltipsConfig.DIVIDER_LINE_BOTTOM_PADDING);
+    }
+
 }